+++ tests/testthat/test-pack.R
@@ -91,6 +91,75 @@ test_that("can unpack 1-row but 0-col dataframe (#1189)", {
   expect_identical(unpack(df, x), tibble::new_tibble(list(), nrow = 1L))
 })
 
+test_that("catches across inner name duplication (#1425)", {
+  df <- tibble(
+    x = tibble(a = 3, b = 4),
+    y = tibble(b = 5),
+    z = tibble(a = 6, b = 6)
+  )
+
+  expect_snapshot(error = TRUE, {
+    unpack(df, c(x, y))
+  })
+  expect_snapshot(error = TRUE, {
+    unpack(df, c(x, y, z))
+  })
+})
+
+test_that("catches outer / inner name duplication (#1367)", {
+  df <- tibble(
+    a = 1,
+    b = 2,
+    c = 3,
+    d = tibble(a = 4),
+    e = tibble(d = 5),
+    f = tibble(b = 6, c = 7, g = 8)
+  )
+
+  expect_snapshot(error = TRUE, {
+    unpack(df, d)
+  })
+  expect_snapshot(error = TRUE, {
+    unpack(df, c(d, e, f))
+  })
+})
+
+test_that("duplication error isn't triggered on the names you are unpacking", {
+  df <- tibble(x = tibble(x = 1))
+  expect_identical(unpack(df, x), tibble(x = 1))
+})
+
+test_that("duplication errors aren't triggered if `names_sep` is specified", {
+  df1 <- tibble(
+    x = 1,
+    y = tibble(x = 2)
+  )
+  df2 <- tibble(
+    x = tibble(a = 1),
+    y = tibble(a = 2)
+  )
+
+  expect_identical(
+    unpack(df1, y, names_sep = "_"),
+    tibble(x = 1, y_x = 2)
+  )
+  expect_identical(
+    unpack(df2, c(x, y), names_sep = "_"),
+    tibble(x_a = 1, y_a = 2)
+  )
+})
+
+test_that("duplication errors aren't triggered on duplicates within a single column you are unpacking", {
+  df <- tibble(
+    x = tibble(a = 1, a = 2, .name_repair = "minimal")
+  )
+
+  # `vec_as_names()` handles this one
+  expect_snapshot(error = TRUE, {
+    unpack(df, x)
+  })
+})
+
 test_that("unpack() validates its inputs", {
   df <- tibble(x = 1:2, y = tibble(a = 1:2, b = 1:2))